Furthering the exceptions strategy to simplify the error response handling

+++ admin-api/controllers/people.php
@@ -2,6 +2,9 @@
 
 use Symfony\Component\HttpFoundation\Request;
 use app\models\People;
+use app\exceptions\ServiceException;
+use app\exceptions\ApiException;
+
 
 $app->get('/person/{id}', function($id, Request $request) use ($app) {
     return $app->json($app['people']->get((int) $id));
@@ -32,7 +35,7 @@ $app->post('/person/{id}', function($id, Request $request) use ($app) {
     $saved = $app['people']->update($id, $data);
 
     if (!$saved) {
-        return $app->json(null, 500, ['X-Error-Message' => 'Unable to save person data']);
+        throw new ServiceException('Unable to save person data');
     }
 
     if ($person['type'] === People::TYPE_GUEST) {
@@ -41,7 +44,7 @@ $app->post('/person/{id}', function($id, Request $request) use ($app) {
             $guest_id = $person['guest_id'];
             $saved = $app['guests']->update($guest_id, compact('food_concerns'));
             if (!$saved) {
-                return $app->json(null, 500, ['X-Error-Message' => 'Unable to save guest data']);
+                throw new ServiceException('Unable to save guest data');
             }
         }
     }
@@ -54,6 +57,9 @@ $app->post('/person/{id}', function($id, Request $request) use ($app) {
 
 $app->delete('/person/{id}', function ($id) use ($app) {
     $person = $app['people']->get((int) $id);
+    if ($person['status'] == People::STATUS_DELETED) {
+        throw new ApiException("Person $id is already deleted");
+    }
     $result = $app['people']->delete($id);
     return $app->json($result);
 });
